Don't "step" reg effects query
After doing some testing, it was determined that stepping through the regeffects
associated with an experiment wasn't necessary. I think it was in the python
version because of memory pressure due to model object hydration. But that's
no longer a concern.

By not stepping through the results we no longer need to order them, which
results in a much faster reg effect query.