Move Options out of data_structures
To facilitate reuse data_structures will be moved into its own crate,
eventually. Options is a structure sepcific to the cov_viz application
and doesn't need to be shared. So this moves Options out of data_structures,
and also combines it with the read_args function for better coherence
between data and functions on that data.