Change Rc->Arc, RefCell->Mutex, fix data structure
Arc/Mutex changes were made to make code thread safe (Rc/RefCell are not).
This thread safety is important for when the data structure code is pulled
into a library for use by the django extension.

The Interval data structure also dropped information -- it needs to keep RegEffectData
per source/target, not all in one blob.